Create class for user-changable values
Values that can be set programmatically but also change depending on
user action are a bit tricky. The implementation was duplicated at
several places and more places are bound to emerge.

In the implementation, the value can programmatically be set via a
stream, and that is merged with the stream of user events, and results
in a output cell for the actual "current" value.