Make SQWidgetBase::geometry a proper value
And add viewportSize for SQListView.

Make geometry setable with a stream and merge that with user changes
(e.g. when the user changes window size and the layout changes, or drags
a splitter or such). It is a bit tricky because we cannot directly send
the new value from the event handlers, because they can be called in a
delayed manner directly from setVisible, which in turn can happen during
sodium evaluation so we may not send. And we cannot reliably set an
initial value either. Also, move and resize events can happen directly
in sequence.

QListView is also funny, in that its resizeEvent usually is triggered
for the viewport(), not the widget itself. Add a separate property for
the viewport size, since geometry should really be the QListView
geometry.